feat: code restructure
