+++ app_v2/src/exceptions.py
